refactor(update): remove leftover adapter discovery
Delete UpdateAdmission and Adapter checkForUpdate so only
UpdateDiscovery owns channel, version, manifest, and result policy.
Keep Windows and package services install-only.
