+++ src/main/java/featurecat/lizzie/update/WindowsUpdateService.java
@@ -1,6 +1,5 @@
 package featurecat.lizzie.update;
 
-import featurecat.lizzie.Lizzie;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
@@ -8,77 +7,22 @@ import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
-import java.util.Optional;
 import org.json.JSONArray;
 import org.json.JSONObject;
 
 public final class WindowsUpdateService {
-  @Deprecated
-  public static final String DEFAULT_MANIFEST_URL =
-      "https://github.com/wimi321/lizzieyzy-next/releases/latest/download/"
-          + "lizzieyzy-next-update-manifest.json";
-
-  public static final String MANIFEST_URL_PROPERTY =
-      UpdateManifestClient.LEGACY_MANIFEST_URL_PROPERTY;
-
-  private final UpdateManifestClient manifestClient;
   private final ResumableDownloader downloader;
-  private final UpdateChannel channel;
 
   public WindowsUpdateService() {
-    this(UpdateChannel.current(), UpdateSource.current());
+    this(new ResumableDownloader());
   }
 
-  public WindowsUpdateService(UpdateChannel channel, UpdateSource source) {
-    this(new UpdateManifestClient(channel, source), new ResumableDownloader(), channel);
-  }
-
-  WindowsUpdateService(UpdateManifestClient manifestClient, ResumableDownloader downloader) {
-    this(manifestClient, downloader, UpdateChannel.STABLE);
-  }
-
-  WindowsUpdateService(
-      UpdateManifestClient manifestClient,
-      ResumableDownloader downloader,
-      UpdateChannel channel) {
-    this.manifestClient = manifestClient;
+  WindowsUpdateService(ResumableDownloader downloader) {
     this.downloader = downloader;
-    this.channel = channel == null ? UpdateChannel.STABLE : channel;
   }
 
   public interface ProgressListener {
     void onProgress(String status, long completedBytes, long totalBytes);
-  }
-
-  public Optional<WindowsUpdatePlan> checkForUpdate() throws IOException {
-    if (!WindowsUpdatePaths.isWindowsRuntime()) {
-      return Optional.empty();
-    }
-    if (!UpdateAdmission.shouldFetch(Lizzie.nextVersion)) {
-      return Optional.empty();
-    }
-    UpdateAdmission.Result admission =
-        UpdateAdmission.evaluateClient(channel, Lizzie.nextVersion, manifestClient);
-    if (admission.kind == UpdateAdmission.Kind.ERROR) {
-      throw new IOException(admission.message);
-    }
-    if (admission.kind == UpdateAdmission.Kind.NO_UPDATE || admission.manifest == null) {
-      return Optional.empty();
-    }
-    WindowsUpdatePaths paths = WindowsUpdatePaths.detect();
-    InstalledUpdateState installed =
-        InstalledUpdateState.read(paths.appDir, Lizzie.nextVersion, paths.flavor);
-    WindowsUpdatePlan plan =
-        WindowsUpdatePlan.create(admission.manifest, installed, Lizzie.nextVersion, paths.flavor);
-    return plan.hasUpdate() ? Optional.of(plan) : Optional.empty();
-  }
-
-  static boolean isStableRelease(UpdateManifest manifest) {
-    return manifest != null && !manifest.prerelease;
-  }
-
-  public UpdateManifest fetchLatestManifest() throws IOException {
-    return manifestClient.fetchLatest().manifest;
   }
 
   public Path downloadAndPrepare(WindowsUpdatePlan plan, ProgressListener listener)
