+++ src/test/java/featurecat/lizzie/update/UpdateManifestClientTest.java
@@ -1,27 +1,42 @@
 package featurecat.lizzie.update;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpServer;
+import featurecat.lizzie.ConfigTestHelper;
+import featurecat.lizzie.Lizzie;
+import featurecat.lizzie.util.NetworkProxy;
+import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.OutputStream;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
+import java.net.ServerSocket;
+import java.net.Socket;
 import java.nio.charset.StandardCharsets;
+import java.nio.file.Path;
 import java.security.KeyPair;
 import java.security.KeyPairGenerator;
 import java.security.Signature;
 import java.util.Base64;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.atomic.AtomicInteger;
 import org.json.JSONObject;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.AfterEach;
-
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
 
 class UpdateManifestClientTest {
+  @TempDir Path tempDir;
+
   @Test
   void rejectsTamperedPrimaryEnvelopeAndUsesSignedMirror() throws Exception {
     KeyPair pair = KeyPairGenerator.getInstance("Ed25519").generateKeyPair();
@@ -70,6 +85,55 @@ class UpdateManifestClientTest {
   void tearDown() {
     System.clearProperty(UpdateManifestClient.ENVELOPE_URLS_PROPERTY);
     System.clearProperty(UpdateManifestClient.LEGACY_MANIFEST_URL_PROPERTY);
+    Lizzie.config = null;
+  }
+
+  @Test
+  void fetchLatestUsesConfiguredProxy() throws Exception {
+    try (OneShotHttp proxy = new OneShotHttp(UpdateManifestTest.validManifest().toString())) {
+      useManualProxy(proxy.port());
+      System.setProperty(
+          UpdateManifestClient.LEGACY_MANIFEST_URL_PROPERTY, "http://example.invalid/update.json");
+
+      UpdateManifest manifest = new UpdateManifestClient().fetchLatest().manifest;
+
+      assertEquals("next-2026-06-12.1", manifest.releaseTag);
+      assertEquals(1, proxy.requests.get());
+      assertTrue(proxy.lastRequestLine.contains("http://example.invalid/update.json"));
+    }
+  }
+
+  @Test
+  void fetchLatestReportsInvalidProxyConfigAsNetworkFailure() {
+    Lizzie.config = ConfigTestHelper.createForTests(tempDir.resolve("config"));
+    Lizzie.config.uiConfig =
+        new JSONObject()
+            .put(NetworkProxy.KEY_PROXY_MODE, NetworkProxy.MODE_MANUAL)
+            .put(NetworkProxy.KEY_PROXY_HOST, " ")
+            .put(NetworkProxy.KEY_PROXY_PORT, 7897);
+    System.setProperty(
+        UpdateManifestClient.LEGACY_MANIFEST_URL_PROPERTY, "http://example.invalid/update.json");
+
+    IOException error =
+        assertThrows(IOException.class, () -> new UpdateManifestClient().fetchLatest());
+
+    assertFalse(error.getMessage().contains("Invalid update manifest"));
+    assertTrue(error.getMessage().contains(NetworkProxy.KEY_PROXY_HOST));
+    assertTrue(error.getMessage().contains("Settings"));
+  }
+
+  @Test
+  void fetchLatestDoesNotFilterPrereleasePayloads() throws Exception {
+    JSONObject prerelease = UpdateManifestTest.validManifest().put("prerelease", true);
+    try (TestServer server = new TestServer()) {
+      server.context("/update.json", exchange -> respond(exchange, prerelease.toString()));
+      System.setProperty(
+          UpdateManifestClient.LEGACY_MANIFEST_URL_PROPERTY, server.url("/update.json"));
+
+      UpdateManifest manifest = new UpdateManifestClient().fetchLatest().manifest;
+
+      assertTrue(manifest.prerelease);
+    }
   }
 
   @Test
@@ -165,6 +229,70 @@ class UpdateManifestClientTest {
     @Override
     public void close() {
       server.stop(0);
+    }
+  }
+
+  private void useManualProxy(int port) {
+    Lizzie.config = ConfigTestHelper.createForTests(tempDir.resolve("config"));
+    Lizzie.config.uiConfig = new JSONObject();
+    Lizzie.config.uiConfig.put(NetworkProxy.KEY_PROXY_MODE, NetworkProxy.MODE_MANUAL);
+    Lizzie.config.uiConfig.put(NetworkProxy.KEY_PROXY_HOST, "127.0.0.1");
+    Lizzie.config.uiConfig.put(NetworkProxy.KEY_PROXY_PORT, port);
+  }
+
+  private static final class OneShotHttp implements AutoCloseable {
+    final AtomicInteger requests = new AtomicInteger();
+    final ServerSocket server;
+    final ExecutorService executor = Executors.newSingleThreadExecutor();
+    final byte[] body;
+    volatile String lastRequestLine = "";
+
+    OneShotHttp(String body) throws IOException {
+      this(body.getBytes(StandardCharsets.UTF_8));
+    }
+
+    OneShotHttp(byte[] body) throws IOException {
+      this.body = body;
+      this.server = new ServerSocket(0, 50, InetAddress.getLoopbackAddress());
+      executor.submit(this::serve);
+    }
+
+    int port() {
+      return server.getLocalPort();
+    }
+
+    private void serve() {
+      try {
+        while (!server.isClosed()) {
+          handle(server.accept());
+        }
+      } catch (IOException ignored) {
+      }
+    }
+
+    private void handle(Socket socket) throws IOException {
+      requests.incrementAndGet();
+      try (socket) {
+        BufferedReader reader =
+            new BufferedReader(
+                new InputStreamReader(socket.getInputStream(), StandardCharsets.US_ASCII));
+        lastRequestLine = reader.readLine();
+        String line;
+        while ((line = reader.readLine()) != null && !line.isEmpty()) {}
+        OutputStream out = socket.getOutputStream();
+        out.write(
+            ("HTTP/1.1 200 OK\r\nContent-Length: "
+                    + body.length
+                    + "\r\nConnection: close\r\n\r\n")
+                .getBytes(StandardCharsets.US_ASCII));
+        out.write(body);
+      }
+    }
+
+    @Override
+    public void close() throws IOException {
+      server.close();
+      executor.shutdownNow();
     }
   }
 }
